Restore python type-only constructs for simplicity
