--- obstore/python/obstore/__sign.py
@@ -1,14 +1,0 @@
-from typing import Literal, TypeAlias
-
-HTTP_METHOD: TypeAlias = Literal[
-    "GET",
-    "PUT",
-    "POST",
-    "HEAD",
-    "PATCH",
-    "TRACE",
-    "DELETE",
-    "OPTIONS",
-    "CONNECT",
-]
-"""Allowed HTTP Methods for signing."""
