+++ obstore/python/obstore/_attributes.pyi
